fix: Make random splitting true by default
Random splitting divides the dataset into training and validation randomly. Given a  dataset or  time-lagged dataset, it will assign the samples to training or validation randomly as opposed to sequentially. For datasets coming from trajectories showing a single transition event or a few this is important. As dividing the dataset sequentially might completely hide the full transition from both sets. Making both the training and evaluation with very different data.