fix: fixed tests
The introduction of a dynamic modification of the seed number used to shuffle the training data has lead to changes in the training results. In addition, tica results are changed in the third decimal place. We are going to change the tica method anyway so I don't want to spend more time on this.