refactor: remove moltype and whole_molecule_selection
moltype was an argument for MOLINFO to determine the type of molecule. Reading the manual I concluded that it shouldn't be needed and will make the config simpler

whole molecule selection is the selection to make whole, i.e. fix the PBC artifacts. By default we will leave 'all' and if needed it will be automated to consider a single entity per molecule in the system